+++ common/src/manager_service.rs
