+++ src/lib.rs
@@ -2,5 +2,4 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod manager_service;
 pub mod service;
